✨Guide Gestion : unicité des relation

+++ src/Controller/Manager/DefaultController.php
@@ -4,6 +4,7 @@ namespace App\Controller\Manager;
 
 use App\Entity\User;
 use App\Entity\Club;
+use App\Repository\MembershipRepository;
 use App\Service\Emailing;
 use App\Repository\UserRepository;
 use Doctrine\ORM\EntityManagerInterface;
@@ -16,22 +17,34 @@ class DefaultController extends AbstractController
 {
     private Emailing $emailing;
     private EntityManagerInterface $entityManager;
+    private MembershipRepository $MembershipRepository;
+    private UserRepository $userRepository;
+
     public function __construct(
         Emailing $emailing,
         EntityManagerInterface $entityManager,
+        MembershipRepository $MembershipRepository,
+        UserRepository $userRepository
     )
     {
         $this->emailing = $emailing;
         $this->entityManager = $entityManager;
+        $this->MembershipRepository = $MembershipRepository;
+        $this->userRepository = $userRepository;
     }
 
 
     #[Route('/', name: 'default_index', methods: ['GET'])]
     public function index(Request $request, UserRepository $userRepository): Response
     {
+        $ManagerId = $this->getUser()->getId();
+        $clubId = $this->getUser()->getClubs()->first()->getId();
+        $vipusers = $this->userRepository->findUserByMembershipId($clubId, $ManagerId);
         return $this->render('manager/default/index.html.twig', [
             'controller_name' => 'DefaultController',
-            'users' => $userRepository->findAll()
+            'users' => $userRepository->findAll(),
+            'vipusers' => $vipusers,
+
 
         ]);
     }
@@ -70,15 +83,28 @@ class DefaultController extends AbstractController
     #[Route(path:'/delete_guide/{id}', name: 'delete_guide', methods: ['GET', 'POST'])]
     public function delete_Guide(Request $request, UserRepository $userRepository, $id): Response
     {
-        $user = $userRepository->find($id);
-        $memberships = $user->getMemberships();
-        foreach ($memberships as $membership) {
-            $this->entityManager->remove($membership);
+        $user = $this->userRepository->find($id);
+        if (!$user) {
+            throw $this->createNotFoundException('The user does not exist');
         }
+
+        $ManagerId = $this->getUser()->getId();
+        $clubId = $this->getUser()->getClubs()->first()->getId();
+        $vipusers = $this->userRepository->findUserByMembershipId($clubId, $ManagerId);
+        //compare $user with vipusers
+        $membership = $this->MembershipRepository->findOneBy(['user' => $user, 'club' => $this->getUser()->getClubs()->first()]);
+        if (!$membership) {
+            $this->addFlash('error', 'This user is not a guide associated to your club');
+            return $this->redirectToRoute('manager_default_index');
+        }
+
+        $this->entityManager->remove($membership);
         $this->entityManager->flush();
+
         $this->addFlash('success', 'Guide deleted from club');
         return $this->redirectToRoute('manager_default_index');
     }
+
 }
